Fixxed Factoralize

+++ factoralize-number.c
@@ -15,14 +15,14 @@
 // factorialize(5);
 int main(void){
     int num = get_int("Number: ");
-    int value = 1;
+    long long value = 1;
     if(num == 1 || num == 0){
-        printf("%i\n", value);
+        printf("%lli\n", value);
     } else {
         for(int i = 1; i < num; i++){
             value *= i;
         }
-        printf("%i\n", num * value);
+        printf("%lli\n", num * value);
     }
 
 }
